+++ src/main/java/com/example/sendowl/repository/BoardRepository.java
@@ -11,6 +11,6 @@ import java.util.List;
 import java.util.Optional;
 
 public interface BoardRepository extends JpaRepository<Board, Long> {
-    List<Board> findByActive(int active);
+    List<Board> findByActive(Long active);
 }
